Check index size during decomp of index_add (#108826)
This partially fixes the `test_index_add_correctness` test (#108181)
when run under inductor: it causes an exception to be raised [here][1]
as expected.

The test as a whole still cannot be made to pass under inductor because
the [last assert][2] still fails, likely due to #108798.

[1]: https://github.com/pytorch/pytorch/blob/dec2b267d4af159bd0a8669a679700f385a1dc98/test/test_torch.py#L6049
[2]: https://github.com/pytorch/pytorch/blob/dec2b267d4af159bd0a8669a679700f385a1dc98/test/test_torch.py#L6051
Pull Request resolved: https://github.com/pytorch/pytorch/pull/108826
Approved by: https://github.com/eellison