[PyTorch] AOTI: use array of constants (#111815)
We continue to allow the user to set clients with a map, but under the hood we use an array of constants.

model_container thought it was OK to hand over the map, assume we just
kept a pointer, and then mutate the map later; I had to fix that. I
hope there aren't other sites that do the same thing...

Differential Revision: [D50111512](https://our.internmc.facebook.com/intern/diff/D50111512/)

Pull Request resolved: https://github.com/pytorch/pytorch/pull/111815
Approved by: https://github.com/jansel, https://github.com/desertfire