Eliminate try-catch block around triton::_triton_bsr_dense_mm_out call. (#112154)
As in the title.

Currently, the try-catch block hides the failures from triton kernel launches that are not related to exceptions that the try-catch block is meant to ignore. When triton kernel launch fails (e.g. due to bugs in triton or lack of resources), ignoring such failures will lead to hard-to-explain/unrelated errors in subsequent code.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/112154
Approved by: https://github.com/amjames, https://github.com/cpuhrsch